Fix warning: implicit conversion changes signedness: 'uint64_t' (aka 'unsigned long long') to 'const int64_t' (aka 'const long long') [-Wsign-conversion]
visp/modules/core/include/visp3/core/vpImageTools_warp.h:273:43: warning: implicit conversion changes signedness: 'uint64_t' (aka 'unsigned long long') to 'const int64_t' (aka 'const long long') [-Wsign-conversion]
  273 |             const int64_t t_1 = precision - t;
      |                           ~~~   ~~~~~~~~~~^~~
visp/modules/core/include/visp3/core/vpImageTools_warp.h:273:45: warning: implicit conversion changes signedness: 'const int64_t' (aka 'const long long') to 'uint64_t' (aka 'unsigned long long') [-Wsign-conversion]
  273 |             const int64_t t_1 = precision - t;
      |                                           ~ ^
visp/modules/core/include/visp3/core/vpImageTools_warp.h:275:43: warning: implicit conversion changes signedness: 'uint64_t' (aka 'unsigned long long') to 'const int64_t' (aka 'const long long') [-Wsign-conversion]
  275 |             const int64_t s_1 = precision - s;
      |                           ~~~   ~~~~~~~~~~^~~
visp/modules/core/include/visp3/core/vpImageTools_warp.h:275:45: warning: implicit conversion changes signedness: 'const int64_t' (aka 'const long long') to 'uint64_t' (aka 'unsigned long long') [-Wsign-conversion]
  275 |             const int64_t s_1 = precision - s;
      |                                           ~ ^

+++ modules/core/include/visp3/core/vpImageTools_warp.h
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2024 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -270,9 +270,9 @@ void vpImageTools::warpLinear(const vpImage<Type> &src, const vpMatrix &T, vpIma
             const int64_t yi_lower = yi_ & (~0xFFFF);
 
             const int64_t t = yi_ - yi_lower;
-            const int64_t t_1 = precision - t;
+            const int64_t t_1 = static_cast<int64_t>(precision) - static_cast<int64_t>(t);
             const int64_t s = xi_ - xi_lower;
-            const int64_t s_1 = precision - s;
+            const int64_t s_1 = static_cast<int64_t>(precision) - static_cast<int64_t>(s);
 
             const int x_ = static_cast<int>(xi_ >> nbits);
             const int y_ = static_cast<int>(yi_ >> nbits);
